move SF helpers outside the CMSSW package, make vectorial friend tree for leptons, add a ROOT macro snippet (probably to be removed later on)
