+++ examples/LeptonEfficiencyCorrector.cc
@@ -1,7 +1,26 @@
-#include <iostream>
+#ifndef PhysicsTools_NanoAOD_LeptonEfficiencyCorrector_h
+#define PhysicsTools_NanoAOD_LeptonEfficiencyCorrector_h
 
-#include <PhysicsTools/NanoAOD/interface/WeightCalculatorFromHistogram.h>
-#include <PhysicsTools/NanoAOD/interface/LeptonEfficiencyCorrector.h>
+#include <iostream>
+#include <string>
+#include <vector>
+#include <TH2.h>
+#include <TFile.h>
+
+#include "PhysicsTools/NanoAOD/examples/WeightCalculatorFromHistogram.cc"
+
+class LeptonEfficiencyCorrector {
+ public:
+  LeptonEfficiencyCorrector() {effmaps_.clear();}
+  LeptonEfficiencyCorrector(std::vector<std::string> files, std::vector<std::string> histos);
+  ~LeptonEfficiencyCorrector() {}
+
+  float getSF(int pdgid, float pt, float eta);
+  float getSFErr(int pdgid, float pt, float eta);
+
+ private:
+  std::vector<TH2F*> effmaps_;
+};
 
 LeptonEfficiencyCorrector:: LeptonEfficiencyCorrector(std::vector<std::string> files, std::vector<std::string> histos) {
   effmaps_.clear();
@@ -46,3 +65,5 @@ float LeptonEfficiencyCorrector::getSFErr(int pdgid, float pt, float eta) {
   }
   return out;
 }
+
+#endif
