+++ examples/WeightCalculatorFromHistogram.cc
@@ -1,5 +1,22 @@
+#ifndef PhysicsTools_NanoAOD_WeightCalculatorFromHistogram_h
+#define PhysicsTools_NanoAOD_WeightCalculatorFromHistogram_h
+
 #include <iostream>
-#include <PhysicsTools/NanoAOD/interface/WeightCalculatorFromHistogram.h>
+#include <TH1.h>
+
+class WeightCalculatorFromHistogram {
+ public:
+  WeightCalculatorFromHistogram() {}
+ WeightCalculatorFromHistogram(TH1 *histogram, bool verbose=false) : histogram_(histogram), verbose_(verbose) {}
+  ~WeightCalculatorFromHistogram() {}
+  
+  float getWeight(float x, float y=0) const;
+  float getWeightErr(float x, float y=0) const;
+  
+ private:
+  TH1* histogram_;
+  bool verbose_;
+};
 
 float WeightCalculatorFromHistogram::getWeight(float x, float y) const {
   if(histogram_==NULL) {
@@ -31,3 +48,4 @@ float WeightCalculatorFromHistogram::getWeightErr(float x, float y) const {
   }
 }
   
+#endif
